+++ parler/tests/__init__.py
@@ -1,7 +1,8 @@
-from .admin import *
-from .model_construction import *
-from .model_attributes import *
-from .model_inheritance import *
-from .forms import *
-from .urls import *
-from .query_count import *
+# Expose for Django 1.5 and below (before DiscoverRunner)
+from .test_admin import *
+from .test_model_construction import *
+from .test_model_attributes import *
+from .test_model_inheritance import *
+from .test_forms import *
+from .test_urls import *
+from .test_query_count import *