+++ parler/tests/testapp/models.py
@@ -120,9 +120,11 @@ class ProxyModel(ProxyBase):
 class DoubleModel(TranslatableModel):
     shared = models.CharField(max_length=200, default='')
 
+
 class DoubleModelTranslations(TranslatedFieldsModel):
     master = models.ForeignKey(DoubleModel, related_name='base_translations')
     l1_title = models.CharField(max_length=200)
+
 
 class DoubleModelMoreTranslations(TranslatedFieldsModel):
     master = models.ForeignKey(DoubleModel, related_name='more_translations')
@@ -137,9 +139,17 @@ class RegularModel(models.Model):
 class CharModel(TranslatableModel):
     id = models.CharField(max_length=45, primary_key=True)
 
+
 class CharModelTranslation(TranslatedFieldsModel):
     master = models.ForeignKey(CharModel)
     tr_title = models.CharField(max_length=200)
+
+
+class ForeignKeyTranslationModel(TranslatableModel):
+    translations = TranslatedFields(
+        translated_foreign = models.ForeignKey('RegularModel'),
+    )
+    shared = models.CharField(max_length=200)
 
 
 # Prevent regression of issue #51:
@@ -166,3 +176,4 @@ class Pen(TranslatableModel, Product):
     translations = TranslatedFields(
         identifier=models.CharField(blank=False, default='', max_length=255)
     )
+